refactor gamelog to be submodule of ddb
