+++ ddb/__init__.py
@@ -1,0 +1,2 @@
+from . import errors, gamelog
+from .client import BeyondClient, BeyondClientBase
