+++ ddb/gamelog/errors.py
